pyproject.toml
[project]
name = "alpaca-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hades", email = "nischalbhattaraipi@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "alpaca-py>=0.43.2",
    "click>=8.3.1",
    "rich>=14.2.0",
    "rich-click>=1.8.3",
]

[project.scripts]
alpaca-cli = "alpaca_cli.cli.main:cli"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
